app/core/security.py
from jose import jwt, JWTError
from fastapi import HTTPException, status

from app.core.config import settings


def verify_jwt(token: str) -> dict:
    """Verify a Supabase JWT and return the decoded payload.

    Raises HTTPException 401 if the token is invalid or expired.
    """
    try:
        payload = jwt.decode(
            token,
            settings.SUPABASE_JWT_SECRET,
            algorithms=["HS256"],
            audience="authenticated",
        )
        return payload
    except JWTError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid or expired token",
            headers={"WWW-Authenticate": "Bearer"},
        )
