requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# Supabase
supabase==2.11.0
python-jose[cryptography]==3.3.0

# AI / LLM
boto3==1.36.4
cachetools==5.5.1

# PDF
weasyprint==63.1
arabic-reshaper==3.0.0
python-bidi==0.6.7
qrcode[pil]==8.0
python-barcode==0.15.1
Pillow==11.1.0

# Security
slowapi==0.1.9
python-multipart==0.0.20

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
moto[bedrock]==5.0.27

# Dev
python-dotenv==1.0.1
